pyproject.toml
[project]
name = "mlb_stats_mcp"
version = "0.1.0"
description = "MLB Stats API MCP"
authors = [
    {name = "Ethan Weisberg", email = "ethaniweisberg@gmail.com"},
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx",
    "python-dotenv",
    "mlb-statsapi",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pre-commit>=4.2.0",
    "matplotlib>=3.10.3",
    "pybaseball>=2.2.7",
    "fastapi>=0.115.12",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.0.292",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "RUF", # Ruff-specific rules
]
ignore = []
extend-exclude = [
    ".venv",
    "__pycache__",
    ".git",
    ".pytest_cache",
]

[tool.ruff.isort]
known-first-party = ["mlb_stats_mcp"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/**" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["mlb_stats_mcp/tests"]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]
